+++ src/applications/vaos/services/utils.js
@@ -5,11 +5,6 @@
 import environment from 'platform/utilities/environment';
 import { apiRequest } from 'platform/utilities/api';
 
-const USE_MOCK_DATA =
-  !window.Cypress &&
-  environment.isLocalhost() &&
-  !environment.API_URL.includes('review.vetsgov');
-
 function vaosFHIRRequest(url, ...options) {
   return apiRequest(`${environment.API_URL}/vaos/v1/${url}`, ...options);
 }
@@ -53,60 +48,16 @@ export function mapToFHIRErrors(errors, title = null) {
 }
 
 /**
- * Makes an api request using the standard vets-website fetch helpers, but runs through the vaos mock handlers first
+ * Makes an api request using the standard vets-website fetch helpers, but prefixes
+ * the url with the current api path
  *
  * @export
  * @param {string} url The url of the api request
  * @param {Object} options The options object passed to the fetch call
  * @param {...*} rest Remaining parameters passed through to apiRequest helper
- * @returns {Promise} Either the promise returned by apiRequest, or a promise with mock data as the result
+ * @returns {Promise} The promise returned by apiRequest
  */
-export async function apiRequestWithMocks(url, options, ...rest) {
-  /* istanbul ignore if  */
-  if (USE_MOCK_DATA) {
-    // This needs to be lazy loaded to keep it out of the main bundle
-    const handlers = (await import(/* webpackChunkName: "mocks" */ './mocks'))
-      .default;
-
-    // find a matching handler by method and path checks
-    const match = handlers.find(handler => {
-      return (
-        options?.method === handler.method &&
-        (typeof handler.path === 'string'
-          ? url.endsWith(handler.path)
-          : handler.path.test(url))
-      );
-    });
-
-    if (match) {
-      // eslint-disable-next-line no-console
-      console.log(
-        `VAOS mock request: ${options?.method || 'GET'} ${url}\n${
-          options?.body ? JSON.stringify(JSON.parse(options.body), null, 2) : ''
-        }`,
-      );
-
-      // Sometimes it's useful to grab ids or other data from the url, so
-      // this passes through matched regex groups
-      let groups = [];
-      if (match.path instanceof RegExp) {
-        groups = match.path.exec(url).slice(1);
-      }
-
-      const response =
-        typeof match.response === 'function'
-          ? match.response(url, {
-              requestData: options?.body ? JSON.parse(options.body) : null,
-              groups,
-            })
-          : match.response;
-
-      return new Promise(resolve => {
-        setTimeout(() => resolve(response), match.delay || 150);
-      });
-    }
-  }
-
+export async function apiRequestWithUrl(url, options, ...rest) {
   return apiRequest(`${environment.API_URL}${url}`, options, ...rest);
 }
